front ok

+++ src/main/java/com/example/ecommerce/repositories/CategoryRepository.java
@@ -8,8 +8,5 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<CategoryEntity, Long>
-
-{
-
+public interface CategoryRepository extends JpaRepository<CategoryEntity, Integer> {
 }
