+++ src/main/java/com/example/ecommerce/repositories/ProductRepository.java
@@ -7,9 +7,8 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ProductRepository extends JpaRepository<ProductEntity, Long> {
+public interface ProductRepository extends JpaRepository<ProductEntity, Integer> {
 
     List<ProductEntity> findByCategoryId(int id);
-
 
 }
